Add components and handler functions to custom hook module; rename, wire in
